% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_patterns.R
\name{all_patterns}
\alias{all_patterns}
\title{Missing data spike-in in various missing data patterns}
\arguments{
\item{X_hat}{Simulated matrix with no missingess (Simulated_matrix output from the simulate() function)}

\item{missfrac_per_var}{Fraction of missingness per variable (Fraction_missingness_per_variable output from the get_data() function)}

\item{assumed_pattern}{Vector of missingess types (must be same length as missingness fraction per variable). If this input is specified, the function will spike in missing datapoints in a MAP pattern as well.}

\item{window}{Window (with default 0.5). This regulates the "extremity" of missingness spike in (larger windows result in more sparse missing data placement whereas smaller windows result in more dense missing data per value - stronger patterns of missingness)}
}
\value{
\item{MCAR_matrix}{Matrix with MCAR pred-defined missingess pattern (default output)}
\item{MAR_matrix}{Matrix with MAR pred-defined missingess pattern (default output)}
\item{MNAR_matrix}{Matrix with MNAR pred-defined missingess pattern (default output)}
\item{MAP_matrix}{Matrix with MAP pred-defined missingess pattern (optional output)}
}
\description{
all_patterns() spikes in missingness using MCAR, MAR, MNAR (default) and MAP (optional) patterns
}
\details{
This function uses the generated simulated matrix and generates missing datapoints in an assumed pattern per variable.
This is achieved by sorting each column and removing the top/bottom of the column (according to the original missingness).
Selection from top or bottom is random.
}
\examples{
all_patterns(simulated$Simulated_matrix, metadata$Fraction_missingness_per_variable, assumed_pattern = c("MAR", "MCAR", "MCAR", "MAR", "MNAR", "MCAR"))
all_patterns(simulated$Simulated_matrix, metadata$Fraction_missingness_per_variable, window = 0.2)

}
