% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_palette.R
\docType{data}
\name{cbfp}
\alias{cbfp}
\title{set consistent ggplot2 theme and colorblind friendly palette}
\format{
An object of class \code{character} of length 7.
}
\usage{
cbfp
}
\description{
set consistent ggplot2 theme and colorblind friendly palette
}
\keyword{datasets}
