% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MltplxExperiment.R
\name{cell_type_counts.MltplxExperiment}
\alias{cell_type_counts.MltplxExperiment}
\title{Get count of cell types in \code{MltplxExperiment}}
\usage{
\method{cell_type_counts}{MltplxExperiment}(mltplx_experiment)
}
\arguments{
\item{mltplx_object}{\code{MltplxExperiment}}
}
\value{
dataframe of cell type counts in boject
}
\description{
Get count of cell types in \code{MltplxExperiment}
}
