% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MltplxObject.R
\name{cell_type_counts.MltplxObject}
\alias{cell_type_counts.MltplxObject}
\title{Get count of cell types in \code{MltplxObject}}
\usage{
\method{cell_type_counts}{MltplxObject}(mltplx_object)
}
\arguments{
\item{mltplx_object}{\code{MltplxObject}}
}
\value{
dataframe of cell type counts in boject
}
\description{
Get count of cell types in \code{MltplxObject}
}
