% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dimple-exports}
\alias{\%>\%}
\alias{dimple-exports}
\title{dimple exported operators and S3 methods}
\description{
The following functions are imported and then re-exported
from the dimple package to avoid listing the magrittr package
as a Depends of dimple
}

