% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MltplxObject.R
\name{dist_to_df.MltplxObject}
\alias{dist_to_df.MltplxObject}
\title{Convert distance matrix to dataframe}
\usage{
\method{dist_to_df}{MltplxObject}(mltplx_object, reduce_symmetric = FALSE)
}
\arguments{
\item{mltplx_object}{object of type \code{MltplxObject}}

\item{reduce_symmetric}{logical, whether to remove equivalent rows}
}
\value{
tibble with dist information
}
\description{
Convert distance matrix to dataframe
}
