% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_mltplx_objects.R
\name{filter_mltplx_objects}
\alias{filter_mltplx_objects}
\title{Filters \code{MltplxExperiment} object based on IDs of slides}
\usage{
filter_mltplx_objects(mltplx_experiment, slide_ids)
}
\arguments{
\item{mltplx_experiment}{\code{MltplxExperiment} object to filter}

\item{slide_ids}{Vector of slide ids to be kept in result \code{MltplxExperiment}}
}
\value{
List of \code{MltplxObject}s with specified ids
}
\description{
Filters \code{MltplxExperiment} object based on IDs of slides
}
