% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_MltplxExp.R
\name{filter_MltplxExp}
\alias{filter_MltplxExp}
\title{Given a MltplxExperiment object and a vector of ids,
produces a new MltplxExperiment object containing only slides whose
ids are in the id vector.}
\usage{
filter_MltplxExp(mltplx_experiment, id_list)
}
\arguments{
\item{mltplx_experiment}{MltplxExperiment object to be filtered}

\item{id_list}{vector of ids to keep in resulting MltplxExperiment object}
}
\value{
New MltplxExperiment object containing only specified slides
}
\description{
Given a MltplxExperiment object and a vector of ids,
produces a new MltplxExperiment object containing only slides whose
ids are in the id vector.
}
