% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridRect.R
\name{GridRect}
\alias{GridRect}
\title{Creates a matrix with an embedded rectangle with the specified intensity.
Note that if the specified rectangle is not fully contained in the matrix,
it will be truncated.}
\usage{
GridRect(m, n, bot_left_corner_x, bot_left_corner_y, width, height, intensity)
}
\arguments{
\item{m}{Number of rows of overall grid}

\item{n}{Number of columns of overall grid}

\item{bot_left_corner_x}{x coordinate of the bottom left corner of the rectangle}

\item{bot_left_corner_y}{y coordinate of the bottom left corner of the rectangle}

\item{width}{width of rectangle}

\item{height}{height of rectangle}

\item{intensity}{Intensity of point pattern on the rectangle}
}
\value{
Intensity matrix of specified dimension with "rectangle" in
specified location.
}
\description{
Creates a matrix with an embedded rectangle with the specified intensity.
Note that if the specified rectangle is not fully contained in the matrix,
it will be truncated.
}
