% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_boxplots.R
\name{group_boxplots}
\alias{group_boxplots}
\title{Box plots of distances between cell type intensities \code{t1} and \code{t2} across
all slides in \code{mltplx_experiment}, grouped by \code{grouping_var}.}
\usage{
group_boxplots(mltplx_experiment, t1, t2, grouping_var = "Group")
}
\arguments{
\item{mltplx_experiment}{\code{MltplxExperiment} object}

\item{t1}{String representing first cell type}

\item{t2}{String representing second cell type}

\item{grouping_var}{Optional; variable in metadata that boxplots should be
grouped by}
}
\description{
Box plots of distances between cell type intensities \code{t1} and \code{t2} across
all slides in \code{mltplx_experiment}, grouped by \code{grouping_var}.
}
