% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello_world.R
\name{hello_world}
\alias{hello_world}
\title{hello_world.R
This is a dummy file for illustration purposes- I'll remove
it for releases. But between this and the tests I've written,
this should offer a reasonable example.}
\usage{
hello_world(name)
}
\arguments{
\item{name}{string- name you want to say hello to}
}
\value{
string
}
\description{
hello_world.R
This is a dummy file for illustration purposes- I'll remove
it for releases. But between this and the tests I've written,
this should offer a reasonable example.
}
