% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp_wrapper.R
\name{imp_wrapper}
\alias{imp_wrapper}
\title{Imputation algoritm wrapper}
\arguments{
\item{rownum}{Number of rows (samples) in the original dataframe (Rows output from the \code{\link{get_data}} function)}

\item{colnum}{Number of rows (variables) in the original dataframe (Columns output from the \code{\link{get_data}} function)}

\item{cormat}{Correlation matrix of the original dataframe (Corr_matrix output from the \code{\link{get_data}} function)}

\item{missfrac_per_var}{Fraction of missingness per variable (Fraction_missingness_per_variable output from the \code{\link{get_data}} function)}

\item{n.iter}{xxx}

\item{assumed_pattern}{Vector of missingess types (must be same length as missingness fraction per variable) (optional parameter)}
}
\value{
\item{Imputation_RMSE}{Number of complete cases (samples with no missing data in any columns)}
\item{Imputation_RMSE_means}{Total number of rows (samples) in the dataframe}
\item{Best_method_MCAR}{Total number of columns (variables) in the dataframe}
\item{Best_method_MAR}{Total number of columns (variables) in the dataframe}
\item{Best_method_MNAR}{Total number of columns (variables) in the dataframe}
\item{Best_method_MAP}{Total number of columns (variables) in the dataframe}
\item{Plot}{Total number of columns (variables) in the dataframe}
}
\description{
\code{\link{imp_wrapper}} tests the imputation accuracy of all missing data imputation algorithms on matrices with various missing data patterns
}
\details{
This function tests the imputation accuracy of the a curated list of missing data imputation algorithms (16 algoritms at the moment) by comparing the
original simulated matrix with no missingess and the imputed matrices generated by the algorithm using the matrices with
MCAR, MAR, MNAR and (optionally) MAP missingness patterns. The function calculates root-mean-square error (RMSE)
between the imputed datapoints and the original datapoints (that were subsequently set to missing) for each missing data imputation algorithm. The function will
automatically detect whether there is a MAP matrix in the list and calculate RMSE for all matrices provided in the list.
}
\examples{
\dontrun{
test_mi(X_hat = simulated$Simulated_matrix, list = miss_list)
}

}
