% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsd.R
\name{jsd}
\alias{jsd}
\title{Computes given two vectors, computes Jensen-Shannon distance between them.
Vectors should be the same length, and may contain zeroes; however, the
distance will only be computed on locations where both vectors are non-zero.
Note that JSD is symmetric, so jsd(v1, v2) = jsd(v2, v1).}
\usage{
jsd(v1, v2, base = 2, normalize = TRUE)
}
\arguments{
\item{v1}{First vector of values}

\item{v2}{Second vector of values}
}
\value{
JSD as double
}
\description{
Computes given two vectors, computes Jensen-Shannon distance between them.
Vectors should be the same length, and may contain zeroes; however, the
distance will only be computed on locations where both vectors are non-zero.
Note that JSD is symmetric, so jsd(v1, v2) = jsd(v2, v1).
}
