% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kld.R
\name{kld}
\alias{kld}
\title{Computes Kullback–Leibler divergence between two vectors of values.}
\usage{
kld(v1, v2, normalize = TRUE)
}
\arguments{
\item{v1}{Vector of probabilities- this represents the P distribution using
Wikipedia's notation.}

\item{v2}{Vector of probabilities- this represents the Q distribution using
Wikipedia's notation.}
}
\value{
KLD as double
}
\description{
Computes Kullback–Leibler divergence between two vectors of values.
}
