% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_dist.R
\name{lm_dist}
\alias{lm_dist}
\title{Linear model of between-group differences in dist}
\usage{
lm_dist(mltplx_experiment, group_factor, slide_ids = NULL, types = NULL)
}
\arguments{
\item{mltplx_experiment}{MltplxExperiment object}

\item{group_factor}{currently only supports binary group_factor. Must be a string for a column name in metadata.}

\item{slide_ids}{vector of slide_ids}

\item{types}{vector of types}
}
\value{
tibble containing all pairwise type comparisons between the groups in the grouping factor, along with inferential statistics
}
\description{
Linear model of between-group differences in dist
}
