% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_dist.R
\name{lm_qdist}
\alias{lm_qdist}
\title{Linear model of between-group differences in quantile dist at specific quantile}
\usage{
lm_qdist(
  mltplx_experiment,
  group_factor,
  interval,
  agg_fun = median,
  covariates = NULL,
  slide_ids = NULL,
  types = NULL
)
}
\arguments{
\item{mltplx_experiment}{MltplxExperiment object}

\item{group_factor}{currently only supports binary group_factor. Must be a string for a column name in metadata.}

\item{interval}{string, which interval to condition on}

\item{agg_fun}{function to aggregate slide-level measurements to patient level. Must have option to remove NA rows via \code{na.rm = T}.}

\item{covariates}{vector of metadata columns to adjust for in linear model}

\item{slide_ids}{vector of slide_ids}

\item{types}{vector of types}
}
\value{
tibble containing all pairwise type comparisons between the groups in the grouping factor, along with inferential statistics
}
\description{
Linear model of between-group differences in quantile dist at specific quantile
}
