% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeDistMat.R
\name{MakeDistMat}
\alias{MakeDistMat}
\title{Converts intensity matrix (i.e. matrix where columns represent cell types,
rows represent locations, and values represent smoothed intensities) into
a distance matrix with specified distance metric.}
\usage{
MakeDistMat(intensity_mat, dist_metric)
}
\arguments{
\item{intensity_mat}{Intensity matrix}

\item{dist_metric}{A distance metric between intensities;  Any function that
takes in two vectors of equal length and produces a scalar}
}
\value{
Distance matrix with rows/columns named appropriately
}
\description{
Converts intensity matrix (i.e. matrix where columns represent cell types,
rows represent locations, and values represent smoothed intensities) into
a distance matrix with specified distance metric.
}
