% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAR.R
\name{MAR}
\alias{MAR}
\title{Missing data spike-in in MAR pattern}
\arguments{
\item{X_hat}{Simulated matrix with no missingess (Simulated_matrix output from the simulate() function)}

\item{missfrac_per_var}{Fraction of missingness per variable (Fraction_missingness_per_variable output from the get_data() function)}

\item{window}{Window (with default 0.5). This regulates the "extremity" of missingness spike in (larger windows result in more sparse missing data placement whereas smaller windows result in more dense missing data per value - stronger patterns of missingness)}
}
\value{
\item{MAR_matrix}{Matrix with MAR pred-defined missingess pattern}
\item{Summary}{Summary of MAR_matrix including number of missing values per variable}
}
\description{
MAR() spikes in missingness using missing-at-random (MAR) pattern
}
\details{
This function uses the generated simulated matrix and generates missing datapoints in an assumed pattern per variable.
This is achieved by sorting each column and removing the top/bottom of the column (according to the original missingness).
Selection from top or bottom is random.
}
\examples{
MAR(simulated$Simulated_matrix, metadata$Fraction_missingness_per_variable)
MAR(simulated$Simulated_matrix, metadata$Fraction_missingness_per_variable, window = 0.2)

}
