% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MltplxDist.R
\name{new_MltplxDist}
\alias{new_MltplxDist}
\title{Creates \code{MltplxDist} object.}
\usage{
new_MltplxDist(
  mltplx_intensity,
  dist_metric,
  .dist_metric_name,
  symmetric = TRUE
)
}
\arguments{
\item{mltplx_intensity}{Intensity matrix (i.e. matrix where columns represent cell types,
rows represent locations, and values represent smoothed intensities)}

\item{dist_metric}{A distance metric between intensities;  Any function that
takes in two vectors of equal length and produces a scalar}

\item{.dist_metric_name}{String representing the name of the distance metric
used}
}
\value{
\code{MltplxDist} object
}
\description{
Creates \code{MltplxDist} object.
}
