% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MltplxExperiment.R
\name{new_MltplxExperiment}
\alias{new_MltplxExperiment}
\title{Creates a new \code{MltplxExperiment} object.}
\usage{
new_MltplxExperiment(
  x,
  y,
  marks,
  slide_id,
  window_sizes = NULL,
  ps = NULL,
  bw = NULL,
  dist_metric = NULL,
  metadata = NULL,
  windows = NULL
)
}
\arguments{
\item{x}{Vector of x coordinates of cells across all slides}

\item{y}{Vector of y coordinates of cells across all slides}

\item{marks}{Vector of cell types for each cell across all slides}

\item{slide_id}{Vector of ids indicating the slide to which each cell
belongs}

\item{window_sizes}{dataframe with columns \code{slide_id}, \code{min_x}, \code{max_x}, \code{min_y} and \code{max_y} indicating the window dimensions for each slide}

\item{ps}{Optional: "pixel size" of intensity estimations. If you'd
like to generate intensities while constructing the \code{MltplxExperiment}
object, it will be broken up into squares that are roughly \code{ps} by \code{ps}
units.}

\item{bw}{Required if \code{ps} is passed. This determines the bandwidth of the
smoothing of the values assigned to each square of the intensity
estimations. Larger values result in "smoother" intensities, while smaller
values result in "coarser" estimations.}

\item{dist_metric}{Optional: a function that computes the distance between
two intensity vectors. This can be any function that takes in two vectors
and returns a scalar.}

\item{metadata}{Optional: patient/slide level data. Must include \emph{one} row
per unique slide id, and a column \code{slide_id} indicating the slide to which each row corresponds}

\item{windows}{Optional: list of \code{owin} objects representing the windows of each
biopsy. Windows will be assigned to slides in the order that they're passed
in accordance with how the slide IDs are ordered.}
}
\value{
\code{MltplxExperiment} object.
}
\description{
This class represents a collection
of multiplex images, which may or may not be from distinct patients. In
addition to storing the images themselves, this class can store estimates
of the intensities of various point types, distances between those
intensities, and metadata related to the patients that the slides belong to.
}
\examples{
set.seed(24601)
cell_x_values = runif(3000, 0, 600)
cell_y_values = runif(3000, 0, 600)
cell_marks = sample(c("Tumor", "Immune", "Other"), 3000, replace = TRUE)
slide_ids = rep(paste("Slide", 1:10), each = 300)

# Create multiplex experiment object
multiplex_experiment = new_MltplxExperiment(x = cell_x_values,
                             y = cell_y_values,
                             marks = factor(cell_marks),
                             slide_id = slide_ids)

# Plot first slide
plot(multiplex_experiment[[1]])

# Update intensities
multiplex_experiment = multiplex_experiment \%>\%
                        update_intensity(ps = 10, bw = 30)

# Plot intensity surfaces of tumor cells and immune cells for slide
# 1 only
plot_intensity_surface(multiplex_experiment,
slide_ids = c("Slide 1"), types = c("Tumor", "Immune"))

# Compute distances between intensities
multiplex_experiment = multiplex_experiment \%>\%
                        update_dist(jsd)

# Plot resulting distances between cell types
plot_dist_matrix(multiplex_experiment, slide_ids = c("Slide 1"))

# Can also be created from a list of `ppp` objects using
`as_MltplxExperiment` function
require(spatstat)
mxp_from_ppp_list = map(1:10, \(i){
spatstat.geom::ppp(
  x = runif(10),
  y = runif(10),
  marks = sample(c("Type 1", "Type 2"), 10, replace = T)
)
}) \%>\%
as_MltplxExperiment()

plot(mxp_from_ppp_list[[1]])
}
