% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MltplxImage.R
\name{new_MltplxImage}
\alias{new_MltplxImage}
\title{Creates new \code{MltplxImage} object.}
\usage{
new_MltplxImage(x, y, marks)
}
\arguments{
\item{x}{X coordinates of cells}

\item{y}{Y coordinates of cells}

\item{marks}{Cell types}
}
\value{
\code{MltplxImage} object
}
\description{
Creates new \code{MltplxImage} object.
}
