% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MltplxIntensity.R
\name{new_MltplxIntensity}
\alias{new_MltplxIntensity}
\title{Creates new \code{MltplxIntensity} object.}
\usage{
new_MltplxIntensity(mltplx_image, ps, bw)
}
\arguments{
\item{mltplx_image}{\code{MltplxImage} object}

\item{ps}{"Pixel size," i.e. the size of the squares that the domain will
be broken up into in order to estimate the intensities of various cell types.}

\item{bw}{Bandwidth- intensity estimates are smoothed, and larger bandwidths
result in more smoothing.}
}
\value{
\code{MltplxIntensity} object
}
\description{
Creates new \code{MltplxIntensity} object.
}
