% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuantileDist.R
\name{new_QuantileDist}
\alias{new_QuantileDist}
\title{Creates a new QuantileDist object from an intensity.}
\usage{
new_QuantileDist(
  mltplx_intensity,
  dist_metric,
  mask_type,
  q_probs,
  dist_metric_name
)
}
\arguments{
\item{mltplx_intensity}{\code{MltplxIntensity} object}

\item{dist_metric}{Distance metric used to evaluate intensity distances}

\item{mask_type}{Name of cell type you want to make a QuantileDist of}

\item{q_probs}{Data frame with columns "from" and "to" with quantile ranges
as percentage points, i.e. 25 for first quartile, 50 for median, etc. Note
that these ranges \emph{can} overlap.}

\item{dist_metric_name}{The distance metric you want to use to compute
distances between distributions.}
}
\value{
object of type \code{QuantileDist}
}
\description{
Creates a new QuantileDist object from an intensity.
}
