% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat.density.R
\name{pat.density}
\alias{pat.density}
\title{Create kernel density from multitype point pattern}
\usage{
pat.density(pat, eps, sigma, ...)
}
\arguments{
\item{pat}{spatstat point pattern}

\item{eps}{pixel size (won't be exactly this size)}

\item{sigma}{function to estimate bandwidth, usually something like bw.diggle}
}
\value{
dataframe in which rows are pixels and columns are cell types
}
\description{
Create kernel density from multitype point pattern
}
