% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MltplxObject.R
\name{plot_dist.MltplxObject}
\alias{plot_dist.MltplxObject}
\title{Plots distance matrix of \code{MltplxObject}, if one has been generated.}
\usage{
\method{plot_dist}{MltplxObject}(mltplx_object, mode = "heatmap")
}
\arguments{
\item{mltplx_object}{Object of class \code{MltplxObject}}

\item{mode}{String indicating plot type, either "heatmap" or "network"}
}
\description{
Plots distance matrix of \code{MltplxObject}, if one has been generated.
}
