% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MltplxObject.R
\name{plot_dist_matrix.MltplxObject}
\alias{plot_dist_matrix.MltplxObject}
\title{Plots distance matrix of \code{MltplxObject}, if one has been generated.}
\usage{
\method{plot_dist_matrix}{MltplxObject}(
  mltplx_object,
  mode = "heatmap",
  net_threshold = 0,
  invert_dist = TRUE,
  symmetric = T
)
}
\arguments{
\item{mltplx_object}{Object of class \code{MltplxObject}}

\item{mode}{String indicating plot type, either "heatmap" or "network"}
}
\description{
Plots distance matrix of \code{MltplxObject}, if one has been generated.
}
