% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dist_regression_heatmap.R
\name{plot_dist_regression_heatmap}
\alias{plot_dist_regression_heatmap}
\title{Pairwise comparison between types across groups}
\usage{
plot_dist_regression_heatmap(df, p_val_col = "p.adj", limits = NULL)
}
\arguments{
\item{df}{tibble. a result of a multiplex analysis, eg, lm_dist}

\item{p_val_col}{name of column with p-values}

\item{limits}{range of values for color scale}
}
\value{
ggplot heatmap
}
\description{
Pairwise comparison between types across groups
}
