% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_intensities.R
\name{plot_intensities}
\alias{plot_intensities}
\title{Plots intensities for selected cell types and selected slides of
\code{MltplxExperiment} object.}
\usage{
plot_intensities(mltplx_experiment, slide_ids, types)
}
\arguments{
\item{mltplx_experiment}{\code{MltplxExperiment} object}

\item{slide_ids}{Vector of slide ids for which you would like to plot
the selected cell intensities.}

\item{types}{Vector of cell types whose intensities you'd like to plot}
}
\description{
Plots intensities for selected cell types and selected slides of
\code{MltplxExperiment} object.
}
