% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MltplxObject.R
\name{plot_intensity_surface.MltplxObject}
\alias{plot_intensity_surface.MltplxObject}
\title{Plots intensities for selected cell types of \code{MltplxObject} object.}
\usage{
\method{plot_intensity_surface}{MltplxObject}(mltplx_object, types = NULL)
}
\arguments{
\item{mltplx_object}{\code{MltplxObject} object}

\item{types}{Vector of cell types whose intensities you'd like to plot}
}
\description{
Plots intensities for selected cell types of \code{MltplxObject} object.
}
