% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pairwise_group_heatmap.R
\name{plot_pairwise_group_heatmap}
\alias{plot_pairwise_group_heatmap}
\title{Pairwise comparison between types across groups}
\usage{
plot_pairwise_group_heatmap(df, group_names = c(1, 2))
}
\arguments{
\item{df}{tibble. a result of a multiplex analysis, eg, lm_dist}
}
\value{
ggplot heatmap
}
\description{
Pairwise comparison between types across groups
}
