% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ppp.R
\name{plot_ppp}
\alias{plot_ppp}
\title{Plots the point processes associated with selected slides of  a
\code{MltplxExperiment} object.}
\usage{
plot_ppp(mltplx_experiment, slide_ids)
}
\arguments{
\item{mltplx_experiment}{\code{MltplxExperiment} object}

\item{slide_ids}{Vector of slide ids whose point processes you would like
to plot.}
}
\description{
Plots the point processes associated with selected slides of  a
\code{MltplxExperiment} object.
}
