% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_qdist.R
\name{plot_qdist}
\alias{plot_qdist}
\title{Plots quantile distance matrices for selected slides in a \code{MltplxExperiment} object}
\usage{
plot_qdist(mltplx_experiment, slide_ids, mode = "heatmap")
}
\arguments{
\item{mltplx_experiment}{\code{MltplxExperiment} objects}

\item{slide_ids}{Vector of ids of slides whose distance matrices you'd like
to print}

\item{mode}{String indicating plot type, either "heatmap" or "network"}
}
\description{
Plots quantile distance matrices for selected slides in a \code{MltplxExperiment} object
}
