% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_qdist_boxplots.R
\name{plot_qdist_boxplots}
\alias{plot_qdist_boxplots}
\title{Plot boxplots of qdist between two groups for a combination of cell types}
\usage{
plot_qdist_boxplots(mltplx_experiment, t1, t2, grouping_var, agg_fun = NULL)
}
\arguments{
\item{t1}{cell type 1}

\item{t2}{cell type 2}

\item{grouping_var}{string, name of metadata column}

\item{agg_fun}{function to aggregate slide-level measurements to patient level.}
}
\value{
ggplot boxplots
}
\description{
Plot boxplots of qdist between two groups for a combination of cell types
}
