% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MltplxObject.R
\name{plot_qdist_matrix.MltplxObject}
\alias{plot_qdist_matrix.MltplxObject}
\title{Plots quantile distance matrices for selected slides in a \code{MltplxExperiment} object}
\usage{
\method{plot_qdist_matrix}{MltplxObject}(mltplx_object, mode = "heatmap", net_threshold = 0)
}
\arguments{
\item{mltplx_object}{\code{MltplxObject} objects}

\item{mode}{String indicating plot type, either "heatmap" or "network"}

\item{net_threshold}{When mode is "network", edges below this absolute value are
excluded from the plot}
}
\value{
ggplot2 object
}
\description{
Plots quantile distance matrices for selected slides in a \code{MltplxExperiment} object
}
