% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_quantile_mask.R
\name{plot_quantile_mask}
\alias{plot_quantile_mask}
\title{Plots quantile of intensities for selected mask cell type and ppp for selected slides of
\code{MltplxExperiment} object.}
\usage{
plot_quantile_mask(mltplx_experiment, mask_type, q_probs, slide_ids)
}
\arguments{
\item{mltplx_experiment}{\code{MltplxExperiment} object}

\item{mask_type}{Name of cell type you want to plot quantiles of intensities}

\item{q_probs}{Data frame with columns "from" and "to" with quantile ranges
as percentage points, i.e. 25 for first quartile, 50 for median, etc. Note
that these ranges \emph{can} overlap.}

\item{slide_ids}{Vector of slide ids for which you would like to plot}
}
\description{
Plots quantile of intensities for selected mask cell type and ppp for selected slides of
\code{MltplxExperiment} object.
}
