% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_simulation_heatmap.R
\name{plot_simulation_heatmap}
\alias{plot_simulation_heatmap}
\title{A custom heatmap function designed to plot matrices as heatmaps.}
\usage{
plot_simulation_heatmap(
  m,
  t = "",
  min.v = NA,
  max.v = NA,
  show_legend = TRUE,
  reflect = FALSE
)
}
\arguments{
\item{m}{Matrix to be plotted}

\item{t}{Optional title of the resulting plot}

\item{min.v}{Minimum value on color legend}

\item{max.v}{Maximum value on color legend}

\item{show_legend}{If TRUE, color legend is shown; if FALSE, it is not.}

\item{reflect}{Whether or not to reflect matrix horizontally. If you're not
sure why you'd want to do this, just leave it as FALSE}
}
\value{
\code{ggplot} object
}
\description{
A custom heatmap function designed to plot matrices as heatmaps.
}
