% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_typewise_comparison.R
\name{plot_typewise_comparison}
\alias{plot_typewise_comparison}
\title{Pairwise comparison between types across groups}
\usage{
plot_typewise_comparison(df)
}
\arguments{
\item{df}{tibble. a result of a multiplex analysis, eg, lm_dist}
}
\value{
ggplot heatmap
}
\description{
Pairwise comparison between types across groups
}
