% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_imp_diag.R
\name{post_imp_diag}
\alias{post_imp_diag}
\title{Post imputation diagnostics}
\usage{
post_imp_diag(X_orig, X_imp, scale = T)
}
\arguments{
\item{X_orig}{Dataframe - the original data that contains missing values.}

\item{X_imp}{Dataframe - the imputed data with no missing values.}

\item{scale}{Boolean with default TRUE. Scaling will scale and center all variables to mean=0 and standard deviation=1 in the \strong{original dataframe with missingness}. The user should select TRUE or FALSE here depending on whether the imputed dataframe has scaled or unscaled values.}
}
\value{
\item{Densityplots}{List of density plots of all variables. The density plots show the original values and the imputed values overlaid for each variables in the dataframe}
\item{Boxplot}{List of boxplots of all variables. The boxplots show the original values and the imputed values for each variables in the dataframe. As normally, the boxplots show the median values, the IQR and the range of values}
}
\description{
\code{\link{post_imp_diag}} serves as post imputation diagnostics. The function compares the original dataset (with missing data) with the imputed dataset. The function outputs visualizations that will help the user compare the distributions of the original values and the imputed values.
}
\details{
This function uses the original dataframe and extracts descriptive meta-data.
}
\examples{
\dontrun{
plots <- post_imp_diag(X_orig = df_miss, X_imp = df_imputed, scale=T)
plots$Densityplots$variable_X
plots$Boxplots$variable_Z
}

}
