% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProgressBar.R
\name{ProgressBar}
\alias{ProgressBar}
\title{Prints text progress bar with spinny thing}
\usage{
ProgressBar(cur, total, ...)
}
\arguments{
\item{cur}{current index}

\item{total}{total number of things to be done}
}
\description{
Prints text progress bar with spinny thing
}
