% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MltplxExperiment.R
\name{qdist_to_df.MltplxExperiment}
\alias{qdist_to_df.MltplxExperiment}
\title{Convert quantile distance matrix to dataframe}
\usage{
\method{qdist_to_df}{MltplxExperiment}(mltplx_experiment, reduce_symmetric = FALSE)
}
\arguments{
\item{mltplx_experiment}{object of type \code{MltplxExperiment}}

\item{reduce_symmetric}{logical, whether to remove equivalent rows}
}
\value{
tibble with dist information
}
\description{
Convert quantile distance matrix to dataframe
}
