% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MltplxObject.R
\name{qdist_to_df.MltplxObject}
\alias{qdist_to_df.MltplxObject}
\title{Quantile distance matrix to dataframe}
\usage{
\method{qdist_to_df}{MltplxObject}(mltplx_object, reduce_symmetric = TRUE)
}
\arguments{
\item{mltplx_object}{object of type \code{MltplxObject}}

\item{reduce_symmetric}{logical, whether to remove equivalent rows}
}
\value{
tibble with dist information
}
\description{
Quantile distance matrix to dataframe
}
