% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MltplxViz.R
\name{sig_stars}
\alias{sig_stars}
\title{GGplot helper to color heatmap tiles by whether or not they are significant according to an adjusted p-value}
\usage{
sig_stars(alpha = 0.05, p_values = "p.adj")
}
\arguments{
\item{alpha}{Significance level}
}
\description{
GGplot helper to color heatmap tiles by whether or not they are significant according to an adjusted p-value
}
