% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateGrid.R
\name{SimulateGrid}
\alias{SimulateGrid}
\title{Generates a simulation of an arbitrary number of cell types with the specified
intensity pattern}
\usage{
SimulateGrid(intensity_list, marks = NULL, square_side_length = 1)
}
\arguments{
\item{intensity_list}{a list of matrices, each of which represents the
intensity pattern of a given cell type. Each matrix should have the same
dimensions.}

\item{square_side_length}{a double indicating the side length of each matrix
entry. Defaults to 1, i.e. each matrix entry is a 1 by 1 unit square.}
}
\description{
Generates a simulation of an arbitrary number of cell types with the specified
intensity pattern
}
