% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_aregImpute.R
\name{test_aregImpute}
\alias{test_aregImpute}
\title{Testing the Hmisc aregImpute missing data imputation algorithm}
\usage{
test_aregImpute(X_hat, list)
}
\arguments{
\item{X_hat}{Simulated matrix with no missingess (this matrix will be used to obtain the error between the original and imputed values). (Simulated_matrix output from the \code{\link{simulate}} function)}

\item{list}{List of matrices with various missingness patterns (MCAR, MAR, MNAR and optionally, MAP). (The input is ideally the R object that was generated using the \code{\link{all_patterns}} function)}
}
\value{
\item{MCAR_RMSE}{Root-mean-square error (RMSE) between the indexed original values and the imputed values in an MCAR missingess pattern (default output)}
\item{MAR_RMSE}{Root-mean-square error (RMSE) between the indexed original values and the imputed values in an MAR missingess pattern (default output)}
\item{MNAR_RMSE}{Root-mean-square error (RMSE) between the indexed original values and the imputed values in an MNAR missingess pattern (default output)}
\item{MAP_RMSE}{Root-mean-square error (RMSE) between the indexed original values and the imputed values in an MAP missingess pattern (optional output)}
}
\description{
\code{\link{test_aregImpute}} tests the imputation accuracy of the Hmisc aregImpute missing data imputation algorithm on matrices with various missing data patterns
}
\details{
This function tests the imputation accuracy of the Hmisc aregImpute missing data imputation algorithm by comparing the
original simulated matrix with no missingess and the imputed matrices generated by the algorithm using the matrices with
MCAR, MAR, MNAR and (optionally) MAP missingness patterns. The function calculates root-mean-square error (RMSE)
between the imputed datapoints and the original datapoints (that were subsequently set to missing). The function will
automatically detect whether there is a MAP matrix in the list and calculate RMSE for all matrices provided in the list.
}
\examples{
\dontrun{
test_aregImpute(X_hat = simulated$Simulated_matrix, list = miss_list)
}

}
