% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pairwise_group_heatmap.R
\name{typewise_boxplots}
\alias{typewise_boxplots}
\title{Plot boxplots of dist between two groups for a combination of cell types}
\usage{
typewise_boxplots(mltplx_experiment, t1, t2, group_factor, agg_fun = NULL)
}
\arguments{
\item{mltplx_experiment}{}

\item{t1}{cell type 1}

\item{t2}{cell type 2}

\item{group_factor}{string, name of metadata column}

\item{agg_fun}{function to aggregate slide-level measurements to patient level.}
}
\value{
ggplot boxplots
}
\description{
Plot boxplots of dist between two groups for a combination of cell types
}
