% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_dist.R
\name{update_dist}
\alias{update_dist}
\title{Update the estimated distances between intensities for all cell types of
all slides of a \code{MltplxExperiment} object, or creates them if there are none.}
\usage{
update_dist(mltplx_experiment, dist_metric, symmetric = TRUE, ...)
}
\arguments{
\item{mltplx_experiment}{The \code{MltplxExperiment} object whose distance
matrices you wish to update}

\item{dist_metric}{The distance metric you'd like to use to compute distance
between vectorized intensities. This can be any function that takes in two
vectors of equal length and returns a scalar.}
}
\value{
Updated \code{MltplxExperiment} object
}
\description{
Update the estimated distances between intensities for all cell types of
all slides of a \code{MltplxExperiment} object, or creates them if there are none.
}
