% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_intensity.R
\name{update_intensity}
\alias{update_intensity}
\title{Update the estimated intensities for all slides of a \code{MltplxExperiment}
object, or creates them if there are none.}
\usage{
update_intensity(mltplx_experiment, ps, bw)
}
\arguments{
\item{mltplx_experiment}{\code{MltplxExperiment} object whose intensities you wish
to update}

\item{ps}{"Pixel size" of the intensity estimations. Each slide will be
broken up into squares of size approximately \code{ps} by \code{ps} units.}

\item{bw}{Bandwidth of the smoothing of the intensity estimations. Larger
values lead to more smoothing. Note that more smoothing is not necessarily
better, and there is no single correct value for smoothing.}
}
\value{
Updated \code{MltplxExperiment} object
}
\description{
Update the estimated intensities for all slides of a \code{MltplxExperiment}
object, or creates them if there are none.
}
