% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_metadata.R
\name{update_metadata}
\alias{update_metadata}
\title{Replaces the metadata of a \code{MltplxExperiment} object with new metadata,
or creates it if there is currently none.}
\usage{
update_metadata(mltplx_experiment, metadata)
}
\arguments{
\item{mltplx_experiment}{\code{MltplxExperiment} object to be updated}

\item{metadata}{New tibble or data.frame of metadata. Must include \emph{one} row
per unique slide id, and a column \code{slide_id} indicating the slide to which
each row corresponds.}
}
\value{
Updated \code{MltplxExperiment} object
}
\description{
Replaces the metadata of a \code{MltplxExperiment} object with new metadata,
or creates it if there is currently none.
}
