% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MltplxObject.R
\name{update_object_dist}
\alias{update_object_dist}
\title{Update the distance matrices generated for a specific \code{MltplxObject}}
\usage{
update_object_dist(
  mltplx_object,
  dist_metric,
  .dist_metric_name = NULL,
  symmetric = TRUE
)
}
\arguments{
\item{mltplx_object}{object of class \code{MltplxObject} to be updated}

\item{dist_metric}{distance metric to be used; can be any function that takes
in two vectors of the same length and produces a scalar}

\item{.dist_metric_name}{Optional; you can use this argument to specify the
name of the distance metric you use}
}
\description{
Update the distance matrices generated for a specific \code{MltplxObject}
}
