% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MltplxObject.R
\name{update_object_intensity}
\alias{update_object_intensity}
\title{Update the intensities generated for a specific \code{MltplxObject}}
\usage{
update_object_intensity(mltplx_object, ps, bw)
}
\arguments{
\item{mltplx_object}{object of class \code{MltplxObject} to be updated}

\item{ps}{"Pixel size" of intensity estimations. Results in squares that are
roughly \code{ps} by \code{ps} units.}

\item{bw}{This determines the bandwidth of the
smoothing of the values assigned to each square of the intensity
estimations. Larger values result in "smoother" intensities, while smaller
values result in "coarser" estimations.}
}
\description{
Update the intensities generated for a specific \code{MltplxObject}
}
