% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MltplxExperiment.R
\name{update_qdist.MltplxExperiment}
\alias{update_qdist.MltplxExperiment}
\title{Update the quantile distance matrices generated for a \code{MltplxExperiment}}
\usage{
\method{update_qdist}{MltplxExperiment}(
  mltplx_experiment,
  dist_metric,
  mask_type,
  q_probs,
  .dist_metric_name = NULL
)
}
\arguments{
\item{mltplx_experiment}{object of class \code{MltplxObject} to be updated}

\item{dist_metric}{distance metric to be used; can be any function that takes
in two vectors of the same length and produces a scalar}

\item{mask_type}{Name of cell type you want to make a QuantileDist of}

\item{q_probs}{Data frame with columns "from" and "to" with quantile ranges
as percentage points, i.e. 25 for first quartile, 50 for median, etc. Note
that these ranges \emph{can} overlap.}

\item{.dist_metric_name}{The distance metric you want to use to compute
distances between distributions.}
}
\value{
updated \code{MltplxExperiment} object
}
\description{
Update the quantile distance matrices generated for a \code{MltplxExperiment}
}
