% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dimple.R
\name{dimple}
\alias{dimple}
\title{dimple}
\usage{
dimple(data, ..., chartType = "bar", stacked = ifelse(chartType \%in\%
  c("bar", "area", "steparea"), T, F), width = 590, height = 400)
}
\arguments{
\item{data}{a data.frame}

\item{...}{other parameters passed to the dimple function (have a look at the examples)}

\item{chartType}{define type of chart, either one of: bar, line, step, area, steparea, pie, bubble}

\item{stacked}{should data be stacked, defaults to TRUE for bar, area and steparea charts.}

\item{width}{width of the chart}

\item{height}{height of the chart}

\item{xPct}{character string of variable to use for a percent x-axis (can be a vector of multiple variables)}

\item{xMeasure}{character string of variable to use for a measure x-axis (can be a vector of multiple variables)}

\item{xCategory}{character string of variable to use for a category x-axis (can be a vector of multiple variables)}

\item{yPct}{character string of variable to use for a percent y-axis (can be a vector of multiple variables)}

\item{yMeasure}{character string of variable to use for a measure y-axis (can be a vector of multiple variables)}

\item{yCategory}{character string of variable to use for a category y-axis (can be a vector of multiple variables)}

\item{legend}{whether to show a legend, defaults to FALSE}
}
\description{
A dimple integration with htmlwidgets for R
}

