% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MltplxObject.R
\name{dist_to_df.MltplxObject}
\alias{dist_to_df.MltplxObject}
\title{Title}
\usage{
\method{dist_to_df}{MltplxObject}(mltplx_object, reduce_symmetric = FALSE)
}
\arguments{
\item{reduce_symmetric}{logical, whether to remove equivalent rows}
}
\value{
tibble with dist information
}
\description{
Title
}
