% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp_wrapper.R
\name{imp_wrapper}
\alias{imp_wrapper}
\title{Imputation algoritm wrapper}
\arguments{
\item{rownum}{Number of rows (samples) in the original dataframe (Rows output from the \code{\link{get_data}} function)}

\item{colnum}{Number of rows (variables) in the original dataframe (Columns output from the \code{\link{get_data}} function)}

\item{cormat}{Correlation matrix of the original dataframe (Corr_matrix output from the \code{\link{get_data}} function)}

\item{missfrac_per_var}{Fraction of missingness per variable (Fraction_missingness_per_variable output from the \code{\link{get_data}} function)}

\item{n.iter}{Number of iterations to perform with default 10.}

\item{assumed_pattern}{Vector of missingess types (must be same length as missingness fraction per variable) (optional parameter)}
}
\value{
\item{Imputation_RMSE}{Raw RMSE values per method, per missingness pattern, per iteration)}
\item{Imputation_RMSE_means}{RMSE summary statistics including mean, standard error and 95\% confidence intervals per method and missingness pattern}
\item{Best_method_MCAR}{The best performing method with MCAR assumption. The method with the lowest mean RMSE calculated from RMSE values from all iterations (default output)}
\item{Best_method_MAR}{The best performing method with MAR assumption. The method with the lowest mean RMSE calculated from RMSE values from all iterations (default output)}
\item{Best_method_MNAR}{The best performing method with MNAR assumption. The method with the lowest mean RMSE calculated from RMSE values from all iterations (default output)}
\item{Best_method_MAP}{The best performing method with MAP assumption. The method with the lowest mean RMSE calculated from RMSE values from all iterations (optional output)}
\item{Plot}{Faceted boxplot of RMSE values per missiness pattern and missing data imputation algorithm}
}
\description{
\code{\link{imp_wrapper}} tests the imputation accuracy of all missing data imputation algorithms on matrices with various missing data patterns
}
\details{
This function tests the imputation accuracy of the a curated list of missing data imputation algorithms (16 algoritms at the moment) by comparing the
original simulated matrix with no missingess and the imputed matrices generated by the algorithm using the matrices with
MCAR, MAR, MNAR and (optionally) MAP missingness patterns. The function calculates root-mean-square error (RMSE)
between the imputed datapoints and the original datapoints (that were subsequently set to missing) for each missing data imputation algorithm. The function will
automatically detect whether there is a MAP matrix in the list and calculate RMSE for all matrices provided in the list.
}
\examples{
\dontrun{
#in case there is no assumed missingness pattern per variable
wrap <- imp_wrapper(rownum = y$Rows,
                    colnum = y$Columns,
                    cormat = y$Corr_matrix,
                    missfrac_per_var =  y$Fraction_missingness_per_variable,
                    n.iter = 15, assumed_pattern = NA)

#in case there is a pre-defined assumed pattern
wrap <- imp_wrapper(rownum = y$Rows,
                    colnum = y$Columns,
                    cormat = y$Corr_matrix,
                    missfrac_per_var =  y$Fraction_missingness_per_variable,
                    n.iter = 15,
                    assumed_pattern = c(rep("MAR", 4), "MCAR", rep("MNAR",2)))
}

}
