% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAP.R
\name{MAP}
\alias{MAP}
\title{Missing data spike-in in MAP pattern}
\arguments{
\item{X_hat}{Simulated matrix with no missingess (Simulated_matrix output from the simulate() function)}

\item{missfrac_per_var}{Fraction of missingness per variable (Fraction_missingness_per_variable output from the get_data() function)}

\item{assumed_pattern}{Vector of missingess types (must be same length as missingness fraction per variable)}

\item{window}{Window (with default 0.5). This regulates the "extremity" of missingness spike in (larger windows result in more sparse missing data placement whereas smaller windows result in more dense missing data per value - stronger patterns of missingness)}
}
\value{
\item{MAP_matrix}{Matrix with MAP pred-defined missingess pattern}
\item{Summary}{Summary of MAP_matrix including number of missing values per variable}
}
\description{
MAP() spikes in missingness using missing-at-assumed (MAP) pattern
}
\details{
This function uses the generated simulated matrix and generates missing datapoints in a missing-not-at-random
pattern for each variable, considering the fraction of missingness for each variable, so potential missing data fraction
imbalances between variables in the original data will be retained. Here the user needs to define a character vector
(of length the same as the fraction of missingness per variable vector) that specifies which missingness pattern corresponds
to the variables. In case the first four columns are assumed missing at random, the next one missing completely at random and
the last two column not at random, the input vector will be:
c(rep("MAR", 4), "MCAR", rep("MNAR",2))
The algorithm will spike in missing values according to the specified pattern. For more information, please see documentation for fuctions
MCAR(), MAR() and MNAR().
}
\examples{
MAP(simulated$Simulated_matrix, metadata$Fraction_missingness_per_variable, assumed_pattern = c("MAR", "MCAR", "MCAR", "MAR", "MNAR", "MCAR"))
MAP(simulated$Simulated_matrix, metadata$Fraction_missingness_per_variable, assumed_pattern = c("MAR", "MCAR", "MCAR", "MAR", "MNAR", "MCAR"), window = 0.2)

}
