% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MNAR.R
\name{MNAR}
\alias{MNAR}
\title{Missing data spike-in in MNAR pattern}
\arguments{
\item{X_hat}{Simulated matrix with no missingess (Simulated_matrix output from the \code{\link{simulate}} function)}

\item{missfrac_per_var}{Fraction of missingness per variable (Fraction_missingness_per_variable output from the \code{\link{get_data}} function)}

\item{window}{Window (with default 0.5). This regulates the "extremity" of missingness spike in (larger windows result in more sparse missing data placement whereas smaller windows result in more dense missing data per value - stronger patterns of missingness)}
}
\value{
\item{MNAR_matrix}{Matrix with MNAR pred-defined missingess pattern}
\item{Summary}{Summary of MNAR_matrix including number of missing values per variable}
}
\description{
\code{\link{MNAR}} spikes in missingness using missing-not-at-random (MNAR) pattern
}
\details{
This function uses the generated simulated matrix and generates missing datapoints in a missing-not-at-random
pattern for each variable, considering the fraction of missingness for each variable, so potential missing data fraction
imbalances between variables in the original data will be retained. The characteristic of the MNAR pattern is that
the missingness in a variable is dependent on its own distribution. Therefore, this function sorts each column (variable)
and based on percentile ranks, randomly selects a window in which missing datapoints will be spiked in. The size of the
window is the fraction of missingness + 0.5 by default. This means that e.g. given 10\% missingness, a window of
10\% + 50\% = 60\% of the data (60 percentiles) will be randomly selected from which 10\% of the variable data
will be randomly removed. In case the user defines a narrower window, the density of missing data removal will be more extreme.
E.g. in an extreme case where the window is defined as 0, 10\% of the variable data will be removed from 10 adjacent
percentiles (for example, the 10\% lowest observations or data between the 72nd and 82nd percentiles). The default size
of the window is 0.5, which means that with this setting the fuction will only run with maximum 50\% of missing data for
any variable. Consequently, with a window setting of 0.25, the function will allow maximum 75\%  missing data a variable.
The user should note that in case the fraction of missingness + the window gets closer to 100\% (e.g. a window of 0.4 and
a variable missingess of 55\%), the more the algorithm will spike in missingness that will resemble a random pattern
(as the random "window" will represent (almost) all the data in a variable). Hence, it is suggested that the user carefully
examines the missing data fractions, excludes variables with high missingess using the \code{\link{clean}} function and
sets a sensible window for the analysis.
}
\examples{
\dontrun{
MNAR(simulated$Simulated_matrix, metadata$Fraction_missingness_per_variable)
MNAR(simulated$Simulated_matrix, metadata$Fraction_missingness_per_variable, window = 0.2)
}

}
