% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MltplxImage.R
\name{new_MltplxImage}
\alias{new_MltplxImage}
\title{Creates new \code{MltplxImage} object.}
\usage{
new_MltplxImage(x, y, marks, xrange = NULL, yrange = NULL, window = NULL)
}
\arguments{
\item{x}{X coordinates of cells}

\item{y}{Y coordinates of cells}

\item{marks}{Cell types}

\item{xrange}{vector of size 2 with range of x-coordinates. If NULL, will default to c(min(x),max(x))}

\item{yrange}{vector of size 2 with range of y-coordinates. If NULL, will default to c(min(y),max(y))}

\item{window}{Optional; object of class \code{owin} from spatstat.geom
representing window of point process}
}
\value{
\code{MltplxImage} object
}
\description{
Creates new \code{MltplxImage} object.
}
