% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient_boxplots.R
\name{patient_boxplots}
\alias{patient_boxplots}
\title{Box plots of distances between cell type intensities \code{t1} and \code{t2} across
all slides in \code{mltplx_experiment}, grouped by patient.}
\usage{
patient_boxplots(
  mltplx_experiment,
  t1,
  t2,
  grouping_var = "Group",
  label_spots = TRUE
)
}
\arguments{
\item{mltplx_experiment}{\code{MltplxExperiment} object}

\item{t1}{First cell type}

\item{t2}{Second cell type}

\item{grouping_var}{Optional; string to indicate variable to use to further
group boxplots.}
}
\value{
ggplot2 plot
}
\description{
Box plots of distances between cell type intensities \code{t1} and \code{t2} across
all slides in \code{mltplx_experiment}, grouped by patient.
}
