% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MltplxObject.R
\name{qdist_to_df.MltplxObject}
\alias{qdist_to_df.MltplxObject}
\title{Quantile dist to df}
\usage{
\method{qdist_to_df}{MltplxObject}(mltplx_object, reduce_symmetric = FALSE)
}
\arguments{
\item{mltplx_object}{}

\item{reduce_symmetric}{logical, whether to remove equivalent rows}
}
\value{
tibble with dist information
}
\description{
Quantile dist to df
}
